requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
requests>=2.31.0
# openai-whisper>=20231117  # uncomment if transcription needed
